Begin updating for BP 1.9, remove BP 1.5 support; merge contirbution from christianwach - see: https://github.com/ddean4040/BP-Group-Hierarchy/pull/1
git-svn-id: http://plugins.svn.wordpress.org/bp-group-hierarchy/trunk@758142 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ bp-group-hierarchy-actions.php
@@ -46,7 +46,7 @@ function bp_group_hierarchy_init() {
  */
 function bp_group_hierarchy_load_components() {
 
-	if( version_compare( bp_get_version(), '1.9', '>=' ) ) {
+	if( version_compare( (float)bp_get_version(), '1.9', '>=' ) ) {
 		// Load BP 1.9+ class -- still using same file until the new one is written
 		require ( dirname( __FILE__ ) . '/bp-group-hierarchy-classes.php' );
 	} else {
